+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added support for manually overriding validation patterns
 - Added lazy loading for postal code examples
 - Added exceptions when no arguments are passed to the validation rules
+- Added PostalCodes facade
 
 ## [v2.1.0](https://github.com/axlon/laravel-postal-code-validation/tree/v2.1.0) (2020-03-04)
 - Added support for Laravel 7